+++ jib-cli/src/test/java/com/google/cloud/tools/jib/cli/jar/JarFilesTest.java
@@ -72,6 +72,7 @@ public class JarFilesTest {
         "11, eclipse-temurin:11-jre",
         "13, eclipse-temurin:17-jre",
         "17, eclipse-temurin:17-jre",
+        "21, eclipse-temurin:21-jre",
       })
   public void testToJibContainer_defaultBaseImage(int javaVersion, String expectedBaseImage)
       throws IOException, InvalidImageReferenceException {